GitHubBrowser/pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "aiohttp>=3.12.12",
    "asyncio>=3.4.3",
    "asyncpg>=0.30.0",
    "beautifulsoup4>=4.13.4",
    "dateparser>=1.2.1",
    "discord-py>=2.5.2",
    "fuzzywuzzy>=0.18.0",
    "googletrans>=4.0.2",
    "lxml>=5.4.0",
    "pillow>=11.2.1",
    "psutil>=7.0.0",
    "pynacl>=1.5.0",
    "python-dotenv>=1.1.0",
    "python-levenshtein>=0.27.1",
    "pytz>=2025.2",
    "qrcode>=8.2",
    "requests>=2.32.4",
    "sqlalchemy>=2.0.41",
    "wikipedia>=1.4.0",
    "yt-dlp>=2025.6.9",
]
